Refactor(mentor): domSelectors.js를 직관적인 이름으로 변경
